+++ examples/callback/main.go
@@ -1,7 +1,8 @@
 // On this example:
-//  - 10 workers will be running forever
 //  - 30 jobs will be enqueued to be processed by the workers
-//	- 1 of the jobs will be enqueued with a callback, which will be invoked just after the job get processed
+//  - 10 workers will be running until the 30 enqueued jobs get processed
+//	- 1 of the jobs will be enqueued plus a callback, which will be invoked just after the job get processed
+//  - there is a callback to enqueue without job's data
 
 package main
 
@@ -40,13 +41,24 @@ func main() {
 	go func() {
 		for i := 0; i < 30; i++ {
 			if i == 15 {
-				pool.AddComplexTask(i, "", func(data interface{}) bool {
+				// enqueue a job + callback
+				// the job will be processed first and later the callback will be invoked
+				pool.AddTaskCallback(i, func(data interface{}) {
 					log.Printf("Callback for: '%v' *******************************", data)
-					return true
 				})
-			} else {
-				pool.AddTask(i)
+				continue
 			}
+
+			if i == 20 {
+				// enqueue a callback
+				// there is no job to process, so the only thing the worker will do is to invoke the callback function
+				pool.AddCallback(func(data interface{}) {
+					log.Println("simple callback function")
+				})
+				continue
+			}
+
+			pool.AddTask(i)
 		}
 
 		// kill all workers after the currently enqueued jobs get processed