+++ pool.go
@@ -61,7 +61,7 @@ const (
 
 	// poolJobData codes
 	poolJobDataCodeRegular            = "regular"
-	poolJobDataCodeComplex            = "complex"
+	poolJobDataCodeCallback           = "callback"
 	poolJobDataCodeLateKillWorker     = "lateKillWorker"
 	poolJobDataCodeLateKillAllWorkers = "lateKillAllWorkers"
 )
@@ -69,12 +69,15 @@ const (
 // PoolFunc defines the function signature to be implemented by the worker's func
 type PoolFunc func(interface{}) bool
 
+// PoolCallback defines the callback function signature
+type PoolCallback func(interface{})
+
 // poolJobData contains the job data && internal pool data
 type poolJobData struct {
 	Code         string
 	JobData      interface{}
 	Category     string
-	CallbackFunc PoolFunc
+	CallbackFunc PoolCallback
 }
 
 type Pool struct {
@@ -434,7 +437,7 @@ func (st *Pool) SetWorkerFunc(fn PoolFunc) {
 // for a free queue slot to enqueue a new job in case the queue is at full capacity.
 //
 // AddTask will return an error if no new tasks could be enqueued at the execution time. No new tasks could be enqueued during
-// a certain amount of time when WaitUntilNSuccesses meet the stop condition.
+// a certain amount of time when WaitUntilNSuccesses meets the stop condition.
 func (st *Pool) AddTask(data interface{}) error {
 	if !st.doNotProcess {
 		// enqueue a regular job
@@ -448,27 +451,52 @@ func (st *Pool) AddTask(data interface{}) error {
 	return errors.New("No new jobs are accepted at this moment")
 }
 
-// AddComplexTask is similar to AddTask allowing the following extra parameters:
-//	- category
-//	- callback ==> function to be invoked just after the job gets processed
+// AddTaskCallback enqueues a job plus a callback function into a FIFO queue (a channel).
 //
-// The parameter for the job's data (required) accepts any type (interface{}).
+// The parameter for the job's data accepts any type (interface{}).
 //
 // Workers (if alive) will be listening to and picking up jobs from this queue. If no workers are alive nor idle,
 // the job will stay in the queue until any worker will be ready to pick it up and start processing it.
 //
-// The queue in which this function enqueues the jobs has a limit (it was set up at pool initialization). It means that AddTask will wait
+// The worker who picks up this job + callback will process the job first and later will invoke the callback function, passing the job's data as a parameter.
+//
+// The queue in which this function enqueues the jobs has a limit (it was set up at pool initialization). It means that AddTaskCallback will wait
 // for a free queue slot to enqueue a new job in case the queue is at full capacity.
 //
-// AddComplexTask will return an error if no new tasks could be enqueued at the execution time. No new tasks could be enqueued during
-// a certain amount of time when WaitUntilNSuccesses meet the stop condition.
-func (st *Pool) AddComplexTask(data interface{}, category string, callbackFn PoolFunc) error {
+// AddTaskCallback will return an error if no new tasks could be enqueued at the execution time. No new tasks could be enqueued during
+// a certain amount of time when WaitUntilNSuccesses meets the stop condition.
+func (st *Pool) AddTaskCallback(data interface{}, callbackFn PoolCallback) error {
 	if !st.doNotProcess {
-		// enqueue a regular job
+		// enqueue a job + callback
 		st.jobsChan <- poolJobData{
-			Code:         poolJobDataCodeComplex,
+			Code:         poolJobDataCodeCallback,
 			JobData:      data,
-			Category:     category,
+			CallbackFunc: callbackFn,
+		}
+		return nil
+	}
+
+	return errors.New("No new jobs are accepted at this moment")
+}
+
+// AddCallback enqueues a callback function into a FIFO queue (a channel).
+//
+// Workers (if alive) will be listening to and picking up jobs from this queue. If no workers are alive nor idle,
+// the job will stay in the queue until any worker will be ready to pick it up and start processing it.
+//
+// The worker who picks up this job will only invoke the callback function, passing nil as a parameter.
+//
+// The queue in which this function enqueues the jobs has a limit (it was set up at pool initialization). It means that AddCallback will wait
+// for a free queue slot to enqueue a new job in case the queue is at full capacity.
+//
+// AddCallback will return an error if no new tasks could be enqueued at the execution time. No new tasks could be enqueued during
+// a certain amount of time when WaitUntilNSuccesses meets the stop condition.
+func (st *Pool) AddCallback(callbackFn PoolCallback) error {
+	if !st.doNotProcess {
+		// enqueue a callback-only job
+		st.jobsChan <- poolJobData{
+			Code:         poolJobDataCodeCallback,
+			JobData:      nil,
 			CallbackFunc: callbackFn,
 		}
 		return nil
@@ -646,19 +674,16 @@ func (st *Pool) workerFunc(n int) {
 						}
 					}
 
-				// complex job:: category ||/&& callback
-				case poolJobDataCodeComplex:
+				// job + callback || callback
+				case poolJobDataCodeCallback:
 					if st.doNotProcess {
-						// TODO ::: re-enqueue in a different queue/channel/struct
-						// re-enqueue the job / task
-						st.AddComplexTask(taskData.JobData, taskData.Category, taskData.CallbackFunc)
-					} else {
+						st.AddTaskCallback(taskData.JobData, taskData.CallbackFunc)
+					}
+
+					// execute the job (in case there is some job data)
+					if taskData.JobData != nil {
 						// execute the job
 						fnSuccess := st.fn(taskData.JobData)
-						if taskData.CallbackFunc != nil {
-							// callback
-							taskData.CallbackFunc(taskData.JobData)
-						}
 
 						// avoid to cause deadlock
 						if !st.doNotProcess {
@@ -668,6 +693,9 @@ func (st *Pool) workerFunc(n int) {
 							// TODO ::: save the job result ...
 						}
 					}
+
+					// run the callback
+					taskData.CallbackFunc(taskData.JobData)
 
 				// late kill signal
 				case poolJobDataCodeLateKillWorker: